Support generation of SVG glyphs using region-flags data.
collect_emoji_svg takes the region-flags data in third_party
and copies/renames it into a target dir, then overlays this with
the data in noto/color_emoji/svg (which also has a few flags in
it).

generate_test_html changes the html to make room for the flag glyphs,
which are a lot wider. It also adds a flag so we can skip regenerating
the .ttx file in case we want to muck with it directly and regenerate
the .woff from that.

svg_builder has a number of changes needed to deal with the much
more varied svg in the region_flags data.  See comments in the file
for some of the changes that needed to be made.  Some of these
changes are provisional pending clarification of the spec, which
doesn't provide enough guidance.

svg_cleaner is changed slightly to strip width and height, this step
now happens after extract this information from svg in svg_builder,
instead of before.