Add/re-enable code to compute ulUnicodeRange values in the OS/2 table.
This was originally disabled in noto_lint, and depended on code that
was never checked in.  This adds the supporting data in opentype_data,
along with utility functions to access it.  It adds functions to
font_data that read and write the OS/2 ulUnicodeRange data values,
and that use the opentype data to compute expected values from the
font's cmap.  The noto_lint data is rewritten to use these functions,
but remains disabled.  Autofix code is added to ensure the OS/2 data
reflects what we believe it should be based on the cmap.

The heuristic is a bit different than the original one.  It's still
a pretty rough heuristic.